requirements.txt
aniso8601==3.0.2
Django==2.1.2
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
Markdown==3.0.1
Pillow==5.3.0
promise==2.2.1
PyJWT==1.6.4
pytz==2018.5
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
